requirements.txt
langgraph
langchain-openai
langchain_anthropic
langchain_google
openai
scipy
numpy
pulp
sympy
graphviz
python-dotenv
pubchempy
scipy
pandas
tabulate